Flyttar shutdown hook-logikken så han er tilgjengeleg frå overalt
