+++ apps/etterlatte-utbetaling/src/main/kotlin/no/nav/etterlatte/Application.kt
@@ -1,6 +1,5 @@
 package no.nav.etterlatte
 
-import no.nav.etterlatte.jobs.addShutdownHook
 import no.nav.etterlatte.libs.common.logging.sikkerlogger
 import no.nav.etterlatte.libs.database.migrate
 import no.nav.etterlatte.utbetaling.config.ApplicationContext
