+++ apps/etterlatte-vedtaksvurdering/src/main/kotlin/Application.kt
@@ -1,6 +1,5 @@
 package no.nav.etterlatte
 
-import no.nav.etterlatte.jobs.addShutdownHook
 import no.nav.etterlatte.libs.common.logging.sikkerLoggOppstartOgAvslutning
 import no.nav.etterlatte.libs.database.migrate
 import no.nav.etterlatte.libs.ktor.initialisering.initEmbeddedServer
